+++ http/client_example_test.go
@@ -91,13 +91,13 @@ func ExampleClient() {
 func makeFrame() (frames.Frame, error) {
 	size := 1027
 	now := time.Now()
-	idata := make([]int, size)
+	idata := make([]int64, size)
 	fdata := make([]float64, size)
 	sdata := make([]string, size)
 	tdata := make([]time.Time, size)
 
 	for i := 0; i < size; i++ {
-		idata[i] = i
+		idata[i] = int64(i)
 		fdata[i] = float64(i)
 		sdata[i] = fmt.Sprintf("val%d", i)
 		tdata[i] = now.Add(time.Duration(i) * time.Second)
